feat: Added method to insert robust password information
A method has been added to the SeguridadPassword class to insert detailed information about the strength of a password into the database. The add_password_info method receives the email, username, and password, then checks the password's security using the PasswordSegura function and proceeds to insert the data into the tm_rob_pass table. Prepared statements are used to prevent SQL injection, and true is returned if the insertion is successful. Otherwise, an error message is logged to the JavaScript console. The PasswordSegura function checks if the password meets minimum security requirements, such as having uppercase letters, lowercase letters, numbers, special characters, and a minimum length of 8 characters.