refactor(js): refactor the way the table is loaded The JavaScript for the history was modified to be loaded with the DataTables library.
Additionally, the JS was updated so that when clicking on `btnDetalleEmergencia`, it retrieves the event after the table is loaded. It was necessary to change how `ev_id` is obtained from the first `tr` in the table.

The click event in the modal map was also updated to correctly get the `ev_id` when the table is loaded.